+++ tools/3d_diagrams.py
@@ -137,7 +137,7 @@ def render_cyl_space(space, resolution, data, c):
         add(space, color.update('hsv', [c1, 1, c2 * 0.005]), x, y, z, c)
 
 
-def render_srgb_cyl_space(space, resolution, data, c):
+def render_srgb_cyl_space(space, resolution, factor, data, c):
     """
     Render the sRGB cylindrical space: HSL, HSV, HWB, etc.
 
@@ -145,13 +145,14 @@ def render_srgb_cyl_space(space, resolution, data, c):
     """
 
     x, y, z = data
+    factor = 100 if space in ('hsluv',) else 1
 
     # Render the cylinder by iterating through the hues and mapping them at the farthest
     # point from the center creating a hollow cylinder. Also, render the top and bottom disc caps.
     color = Color("srgb", [])
     for c1, t in itertools.product(
         ((x / resolution) * 360 for x in range(0, resolution + 1)),
-        (((x / resolution), i) for i, x in enumerate(range(0, resolution + 1), 0))
+        (((x / resolution) * factor, i) for i, x in enumerate(range(0, resolution + 1), 0))
     ):
 
         # Offset the plot on every other iteration blend the rows into a mesh
@@ -163,8 +164,8 @@ def render_srgb_cyl_space(space, resolution, data, c):
         # Top disc
         x.append(c2 * math.sin(math.radians(c1)))
         y.append(c2 * math.cos(math.radians(c1)))
-        z.append(1)
-        c.append(color.update(space, [c1, c2, 1]).to_string(hex=True))
+        z.append(factor)
+        c.append(color.update(space, [c1, c2, factor]).to_string(hex=True))
 
         # Bottom disc
         x.append(c2 * math.sin(math.radians(c1)))
@@ -173,10 +174,10 @@ def render_srgb_cyl_space(space, resolution, data, c):
         c.append(color.update(space, [c1, c2, 0]).to_string(hex=True))
 
         # Cylinder portion
-        x.append(1 * math.sin(math.radians(c1)))
-        y.append(1 * math.cos(math.radians(c1)))
+        x.append(factor * math.sin(math.radians(c1)))
+        y.append(factor * math.cos(math.radians(c1)))
         z.append(c2)
-        c.append(color.update(space, [c1, 1, c2]).to_string(hex=True))
+        c.append(color.update(space, [c1, factor, c2]).to_string(hex=True))
 
 
 def plot_space_in_srgb(space, title="", dark=False, resolution=70):
@@ -220,7 +221,7 @@ def plot_space_in_srgb(space, title="", dark=False, resolution=70):
     # Render the space
     if is_srgb_cyl:
         # Render a sRGB cylinder style plot
-        render_srgb_cyl_space(space, resolution, data, c)
+        render_srgb_cyl_space(space, resolution, 1, data, c)
     elif is_labish or is_cyl:
         # Render cylindrical spaces. Lab like spaces are cylindrical,
         # just represented in the Cartesian coordinate system.